working on debug of L_2 Residual
