+++ main.c
@@ -7,32 +7,141 @@ domain, using a parallel unstructured mesh (DMPLEX) to discretize it.\n\n\n";
 #include "mesh.h"
 #include "parameters.h"
 
+
+static PetscErrorCode SetInitialConditions(TS ts, Vec u) {
+    DM dm;
+    PetscReal t;
+    PetscErrorCode ierr;
+
+    PetscFunctionBegin;
+    ierr = TSGetDM(ts, &dm);
+    CHKERRQ(ierr);
+    ierr = TSGetTime(ts, &t);
+    CHKERRQ(ierr);
+    ierr = DMComputeExactSolution(dm, t, u, NULL);
+    CHKERRQ(ierr);
+    ierr = RemoveDiscretePressureNullspace(dm, u);
+    CHKERRQ(ierr);
+    PetscFunctionReturn(0);
+}
+
+static PetscErrorCode MonitorError(TS ts, PetscInt step, PetscReal crtime, Vec u, void *ctx) {
+    PetscErrorCode (*exactFuncs[3])(PetscInt dim, PetscReal time, const PetscReal x[], PetscInt Nf, PetscScalar *u, void *ctx);
+    void *ctxs[3];
+    DM dm;
+    PetscDS ds;
+    Vec v;
+    PetscReal ferrors[3];
+    PetscInt f;
+    PetscErrorCode ierr;
+
+    PetscFunctionBeginUser;
+    ierr = TSGetDM(ts, &dm);
+    CHKERRQ(ierr);
+    ierr = DMGetDS(dm, &ds);
+    CHKERRQ(ierr);
+
+    for (f = 0; f < 3; ++f) {
+        ierr = PetscDSGetExactSolution(ds, f, &exactFuncs[f], &ctxs[f]);
+        CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    }
+    ierr = DMComputeL2FieldDiff(dm, crtime, exactFuncs, ctxs, u, ferrors);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = PetscPrintf(PETSC_COMM_WORLD, "Timestep: %04d time = %-8.4g \t L_2 Error: [%2.3g, %2.3g, %2.3g]\n", (int)step, (double)crtime, (double)ferrors[0], (double)ferrors[1], (double)ferrors[2]);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+
+    ierr = DMGetGlobalVector(dm, &u);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    // ierr = TSGetSolution(ts, &u);CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = PetscObjectSetName((PetscObject)u, "Numerical Solution");
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = VecViewFromOptions(u, NULL, "-sol_vec_view");
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = DMRestoreGlobalVector(dm, &u);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+
+    ierr = DMGetGlobalVector(dm, &v);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    // ierr = VecSet(v, 0.0);CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = DMProjectFunction(dm, 0.0, exactFuncs, ctxs, INSERT_ALL_VALUES, v);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = PetscObjectSetName((PetscObject)v, "Exact Solution");
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = VecViewFromOptions(v, NULL, "-exact_vec_view");
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+    ierr = DMRestoreGlobalVector(dm, &v);
+    CHKERRABORT(PETSC_COMM_WORLD, ierr);
+
+    PetscFunctionReturn(0);
+}
+
+/*
+  CASE: quadratic
+  In 2D we use exact solution:
+
+
+*/
 static PetscErrorCode quadratic_u(PetscInt Dim, PetscReal time, const PetscReal X[], PetscInt Nf, PetscScalar *u, void *ctx) {
-  u[0] = time + X[0] * X[0] + X[1] * X[1];
-  u[1] = time + 2.0 * X[0] * X[0] - 2.0 * X[0] * X[1];
-  return 0;
+    // u = {t + x^2 + y^2, t + 2*x^2 + 2*x*y}
+    u[0] = time + X[0] * X[0] + X[1] * X[1];
+    u[1] = time + 2.0 * X[0] * X[0] + 2.0 * X[0] * X[1];
+    return 0;
 }
 static PetscErrorCode quadratic_u_t(PetscInt Dim, PetscReal time, const PetscReal X[], PetscInt Nf, PetscScalar *u, void *ctx) {
-  u[0] = 1.0;
-  u[1] = 1.0;
-  return 0;
+    u[0] = 1.0;
+    u[1] = 1.0;
+    return 0;
 }
 
 static PetscErrorCode quadratic_p(PetscInt Dim, PetscReal time, const PetscReal X[], PetscInt Nf, PetscScalar *p, void *ctx) {
-  p[0] = X[0] + X[1] - 1.0;
-  return 0;
+    // p = x + y - 1
+    p[0] = X[0] + X[1] - 1.0;
+    return 0;
 }
 
 static PetscErrorCode quadratic_T(PetscInt Dim, PetscReal time, const PetscReal X[], PetscInt Nf, PetscScalar *T, void *ctx) {
-  T[0] = time + X[0] + X[1];
-  return 0;
+    // T = t + x + y
+    T[0] = time + X[0] + X[1];
+    return 0;
 }
 static PetscErrorCode quadratic_T_t(PetscInt Dim, PetscReal time, const PetscReal X[], PetscInt Nf, PetscScalar *T, void *ctx) {
-  T[0] = 1.0;
-  return 0;
+    T[0] = 1.0;
+    return 0;
 }
 
-/* f0_v = du/dt - f */
+static PetscReal Power(PetscReal x, PetscInt exp){
+    return PetscPowReal(x, exp);
+}
+
+static void f0_quadratic_q(PetscInt dim,
+                           PetscInt Nf,
+                           PetscInt NfAux,
+                           const PetscInt uOff[],
+                           const PetscInt uOff_x[],
+                           const PetscScalar u[],
+                           const PetscScalar u_t[],
+                           const PetscScalar u_x[],
+                           const PetscInt aOff[],
+                           const PetscInt aOff_x[],
+                           const PetscScalar a[],
+                           const PetscScalar a_t[],
+                           const PetscScalar a_x[],
+                           PetscReal t,
+                           const PetscReal X[],
+                           PetscInt numConstants,
+                           const PetscScalar constants[],
+                           PetscScalar f0[]){
+    QIntegrandTestFunction(dim, Nf, NfAux, uOff, uOff_x, u, u_t, u_x, aOff, aOff_x, a, a_t, a_x, t, X, numConstants, constants, f0);
+    const PetscReal S = constants[STROUHAL];
+    const PetscReal Pth = constants[PTH];
+    const PetscReal x = X[0];
+    const PetscReal y = X[1];
+
+    f0[0] -= (Pth*S)/Power(t + x + y,2) - (4*Pth*x)/(t + x + y) +
+             (Pth*(t + 2*Power(x,2) + 2*x*y))/Power(t + x + y,2) +
+             (Pth*(t + Power(x,2) + Power(y,2)))/Power(t + x + y,2);
+}
+
 static void f0_quadratic_v(PetscInt dim,
                            PetscInt Nf,
                            PetscInt NfAux,
@@ -51,12 +160,22 @@ static void f0_quadratic_v(PetscInt dim,
                            PetscInt numConstants,
                            const PetscScalar constants[],
                            PetscScalar f0[]) {
-  VIntegrandTestFunction(dim, Nf, NfAux, uOff, uOff_x, u, u_t, u_x, aOff, aOff_x, a, a_t, a_x, t, X, numConstants, constants, f0);
-    //TODO: fix
-  //  const PetscReal nu = PetscRealPart(constants[NU]);
-//
-//  f0[0] -= (t * (2 * X[0] + 2 * X[1]) + 2 * X[0] * X[0] * X[0] + 4 * X[0] * X[0] * X[1] - 2 * X[0] * X[1] * X[1] - 4.0 * nu + 2);
-//  f0[1] -= (t * (2 * X[0] - 2 * X[1]) + 4 * X[0] * X[1] * X[1] + 2 * X[0] * X[0] * X[1] - 2 * X[1] * X[1] * X[1] - 4.0 * nu + 2);
+    VIntegrandTestFunction(dim, Nf, NfAux, uOff, uOff_x, u, u_t, u_x, aOff, aOff_x, a, a_t, a_x, t, X, numConstants, constants, f0);
+
+    const PetscReal S = constants[STROUHAL];
+    const PetscReal Pth = constants[PTH];
+    const PetscReal mu = constants[MU];
+    const PetscReal R = constants[REYNOLDS];
+    const PetscReal F = constants[FROUDE];
+
+    const PetscReal x = X[0];
+    const PetscReal y = X[1];
+
+    f0[0] -= 1 - (5.333333333333334*mu)/R +
+             (Pth*(S + (4*x + 2*y)*(t + 2*Power(x,2) + 2*x*y) + 2*x*(t + Power(x,2) + Power(y,2))))/(t + x + y);
+
+    f0[1] -= 1 - (4.*mu)/R + Pth/(F*(t + x + y)) +
+             (Pth*(S + 2*x*(t + 2*Power(x,2) + 2*x*y) + 2*y*(t + Power(x,2) + Power(y,2))))/ (t + x + y);
 }
 
 /* f0_w = dT/dt + u.grad(T) - Q */
@@ -78,58 +197,19 @@ static void f0_quadratic_w(PetscInt dim,
                            PetscInt numConstants,
                            const PetscScalar constants[],
                            PetscScalar f0[]) {
-  WIntegrandTestFunction(dim, Nf, NfAux, uOff, uOff_x, u, u_t, u_x, aOff, aOff_x, a, a_t, a_x, t, X, numConstants, constants, f0);
-  f0[0] += -(2 * t + 1 + 3 * X[0] * X[0] - 2 * X[0] * X[1] + X[1] * X[1]);
+    WIntegrandTestFunction(dim, Nf, NfAux, uOff, uOff_x, u, u_t, u_x, aOff, aOff_x, a, a_t, a_x, t, X, numConstants, constants, f0);
+
+    const PetscReal S = constants[STROUHAL];
+    const PetscReal Pth = constants[PTH];
+    const PetscReal Cp = constants[CP];
+    const PetscReal H = constants[HEATRELEASE];
+
+    const PetscReal x = X[0];
+    const PetscReal y = X[1];
+
+    f0[0] -= (Cp*Pth*(S + 2*t + 3*Power(x,2) + 2*x*y + Power(y,2)))/(H*S*(t + x + y));
 }
 
-static PetscErrorCode SetInitialConditions(TS ts, Vec u) {
-  DM dm;
-  PetscReal t;
-  PetscErrorCode ierr;
-
-  PetscFunctionBegin;
-  ierr = TSGetDM(ts, &dm);CHKERRQ(ierr);
-  ierr = TSGetTime(ts, &t);CHKERRQ(ierr);
-  ierr = DMComputeExactSolution(dm, t, u, NULL);CHKERRQ(ierr);
-  ierr = RemoveDiscretePressureNullspace(dm, u);CHKERRQ(ierr);
-  PetscFunctionReturn(0);
-}
-
-static PetscErrorCode MonitorError(TS ts, PetscInt step, PetscReal crtime, Vec u, void *ctx) {
-  PetscErrorCode (*exactFuncs[3])(PetscInt dim, PetscReal time, const PetscReal x[], PetscInt Nf, PetscScalar *u, void *ctx);
-  void *ctxs[3];
-  DM dm;
-  PetscDS ds;
-  Vec v;
-  PetscReal ferrors[3];
-  PetscInt f;
-  PetscErrorCode ierr;
-
-  PetscFunctionBeginUser;
-  ierr = TSGetDM(ts, &dm);CHKERRQ(ierr);
-  ierr = DMGetDS(dm, &ds);CHKERRQ(ierr);
-
-  for (f = 0; f < 3; ++f) {
-    ierr = PetscDSGetExactSolution(ds, f, &exactFuncs[f], &ctxs[f]);CHKERRQ(ierr);
-  }
-  ierr = DMComputeL2FieldDiff(dm, crtime, exactFuncs, ctxs, u, ferrors);CHKERRQ(ierr);
-  ierr = PetscPrintf(PETSC_COMM_WORLD, "Timestep: %04d time = %-8.4g \t L_2 Error: [%2.3g, %2.3g, %2.3g]\n", (int)step, (double)crtime, (double)ferrors[0], (double)ferrors[1], (double)ferrors[2]);CHKERRQ(ierr);
-
-  ierr = DMGetGlobalVector(dm, &u);CHKERRQ(ierr);
-  // ierr = TSGetSolution(ts, &u);CHKERRQ(ierr);
-  ierr = PetscObjectSetName((PetscObject)u, "Numerical Solution");CHKERRQ(ierr);
-  ierr = VecViewFromOptions(u, NULL, "-sol_vec_view");CHKERRQ(ierr);
-  ierr = DMRestoreGlobalVector(dm, &u);CHKERRQ(ierr);
-
-  ierr = DMGetGlobalVector(dm, &v);CHKERRQ(ierr);
-  // ierr = VecSet(v, 0.0);CHKERRQ(ierr);
-  ierr = DMProjectFunction(dm, 0.0, exactFuncs, ctxs, INSERT_ALL_VALUES, v);CHKERRQ(ierr);
-  ierr = PetscObjectSetName((PetscObject)v, "Exact Solution");CHKERRQ(ierr);
-  ierr = VecViewFromOptions(v, NULL, "-exact_vec_view");CHKERRQ(ierr);
-  ierr = DMRestoreGlobalVector(dm, &v);CHKERRQ(ierr);
-
-  PetscFunctionReturn(0);
-}
 
 // extern int ex76(int argc,char **args);
 // int main(int argc,char **args){
@@ -150,7 +230,7 @@ int main(int argc, char **args) {
 
   // setup and initialize the constant field variables
   ierr = PetscBagCreate(PETSC_COMM_WORLD, sizeof(FlowParameters), &context.parameters);CHKERRQ(ierr);
-//  ierr = SetupParameters(&context);CHKERRQ(ierr);//TODO:fix
+  ierr = SetupFlowParameters(&context.parameters);CHKERRQ(ierr);
 
   // setup the ts
   ierr = TSCreate(PETSC_COMM_WORLD, &ts);CHKERRQ(ierr);
@@ -169,7 +249,8 @@ int main(int argc, char **args) {
     ierr = DMGetDS(dm, &prob);CHKERRQ(ierr);
 
     // V, W Test Function
-    ierr = PetscDSSetResidual(prob, V, f0_quadratic_v, VIntegrandTestGradientFunction);CHKERRQ(ierr);
+      ierr = PetscDSSetResidual(prob, Q, f0_quadratic_q, NULL);CHKERRQ(ierr);
+      ierr = PetscDSSetResidual(prob, V, f0_quadratic_v, VIntegrandTestGradientFunction);CHKERRQ(ierr);
     ierr = PetscDSSetResidual(prob, W, f0_quadratic_w, WIntegrandTestGradientFunction);CHKERRQ(ierr);
 
     FlowParameters *parameters;
@@ -230,3 +311,8 @@ int main(int argc, char **args) {
   ierr = PetscFinalize();
   return ierr;
 }
+
+/*
+ * -dm_plex_separate_marker -dm_refine 0  -vel_petscspace_degree 2 -pres_petscspace_degree 1 -temp_petscspace_degree 1  -dmts_check .001 -ts_max_steps 4 -ts_dt 0.1  -ksp_type fgmres -ksp_gmres_restart 10 -ksp_rtol 1.0e-9 -ksp_error_if_not_converged  -pc_type fieldsplit -pc_fieldsplit_0_fields 0,2 -pc_fieldsplit_1_fields 1 -pc_fieldsplit_type schur -pc_fieldsplit_schur_factorization_type full  -fieldsplit_0_pc_type lu  -fieldsplit_pressure_ksp_rtol 1e-10 -fieldsplit_pressure_pc_type jacobi  -ts_fd_color
+ *
+ */
